openim_deploy/backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
httpx==0.27.2
python-jose==3.3.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
